Alteração no método de login e de exclusão de parceiros

+++ src/main/python/views/central_parceiros/parceiros.py
@@ -1,5 +1,12 @@
 from webapp import db, cp, mail
 from models.table_parceiros import Parceiros
+from models.table_inscricoes import Inscricoes
+from models.table_alunos import Alunos
+from models.table_agentes import Agentes
+from models.table_diretores import Diretores
+from models.table_atividades import Atividades
+from models.table_mensagens import Mensagens
+from models.table_avaliacoes import Avaliacoes
 from flask import request, jsonify, url_for
 from werkzeug.security import generate_password_hash
 from views.central_parceiros.login import token_required
@@ -11,7 +18,7 @@ s = URLSafeTimedSerializer('this-is-secret') #melhorar essa chave de segurança
 @cp.route('/parceiro', methods=['GET'])
 @token_required
 def get_parceiro(current_user):
-    dados = Parceiros.query.all()
+    dados = Parceiros.query.alll()
 
     parceiros = []
 
@@ -84,9 +91,9 @@ def post_parceiro():
     db.session.add(parceiro)
     db.session.commit()
     
-    return send_email_confirm(parceiro.email)
+    # return send_email_confirm(parceiro.email)
 
-    #return jsonify({'Mensagem': 'Adicionado com sucesso!'})
+    return jsonify({'Mensagem': 'Adicionado com sucesso!'})
 
 @cp.route('/parceiro/<parceiro_id>', methods=['PUT'])
 @token_required
@@ -159,41 +166,49 @@ def edit_parceiro(current_user, parceiro_id):
 @token_required
 def del_parceiro(current_user, parceiro_id):
     parceiro = Parceiros.query.filter_by(id_geral=parceiro_id).first()
-    inscricao = Inscricoes.query.filter_by(id_parceiros = parceiro_id).first()
 
     if parceiro.nivel == "Aluno":
-        aluno = Alunos.query.filter_by(id_parceiros = parceiro.id_geral).first()
+        aluno = Alunos.query.filter_by(id_parceiros = parceiro_id).first()
 
         db.session.delete(aluno)
         db.session.commit()
 
     elif parceiro.nivel == "Diretor":
-        diretor = Diretores.query.filter_by(id_parceiros = parceiro.id_geral).first()
-        mensagem = Mensagens.query.filter_by(id_destinatario = diretor.id).first()
-
-        db.session.delete(mensagem)
-        db.session.commit()
+        diretor = Diretores.query.filter_by(id_parceiros = parceiro_id).first()
         
         db.session.delete(diretor)
         db.session.commit()
 
     elif parceiro.nivel == 'Agente':
-        agente = Agentes.query.filter_by(id_parceiros = parceiro.id_geral).first()
+        agente = Agentes.query.filter_by(id_parceiros = parceiro_id).first()
 
-        mensagem = Mensagens.query.filter_by(id_destinatarios = parceiro.id_geral).first()
+        db.session.delete(agente)
+        db.session.commit()
 
-        atividades = Atividades.query.filter_by(id_agente = agente.id).first()
 
-        if not atividades and not mensagem:
-
-            db.session.delete(agente)
+    atividades = Atividades.query.filter_by(id_parceiro = parceiro_id).all()
+    if atividades:
+        for atividade in atividades:
+            db.session.delete(atividade)
             db.session.commit()
 
-        else:
-            return jsonify({'Mensagem': 'Você tem algumas atividades pendentes!'})
+    mensagens = Mensagens.query.filter_by(id_destinatarios = parceiro_id).all()
+    if mensagens:
+        for mensagem in mensagens:
+            db.session.delete(mensagem)
+            db.session.commit()
 
-    db.session.delete(inscricao)
-    db.session.commit()
+    inscricoes = Inscricoes.query.filter_by(id_parceiros = parceiro_id).all()
+    if inscricoes:
+        for inscricao in inscricoes:
+            db.session.delete(inscricao)
+            db.session.commit()
+
+    avaliacoes = Avaliacoes.query.filter_by(id_parceiro = parceiro_id).all()
+    if avaliacoes:
+        for avaliacao in avaliacoes:
+            db.session.delete(avaliacao)
+            db.session.commit()
     
     db.session.delete(parceiro)
     db.session.commit()
